pyproject.toml
[project]
name = "todos-from-scratch-2"
version = "0.1.0"
description = ""
authors = [
    {name = "Robert Tsui",email = "rt.roberttsui@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
